+++ amfiteatr_core/src/env/generic/hashmap_environment.rs
@@ -346,7 +346,7 @@ where Self: CommunicatingEndpointEnvironment<DP, CommunicationError=Communicatio
 + EnvironmentWithAgents<DP>
 + BroadcastingEndpointEnvironment<DP>, DP: DomainParameters {
     fn run_truncating(&mut self, truncate_steps: Option<usize>) -> Result<(), AmfiteatrError<DP>> {
-        self.run_round_robin_truncating(truncate_steps)
+        self.run_round_robin_no_rewards_truncating(truncate_steps)
     }
 }